+++ kaizo/utils/__init__.py
@@ -1,6 +1,8 @@
 from .common import extract_variable
 from .entry import DictEntry, Entry, FieldEntry, ListEntry, ModuleEntry
 from .fn import FnWithKwargs
+from .module import ModuleLoader
+from .storage import Storage
 
 __all__ = (
     "DictEntry",
@@ -9,5 +11,7 @@ __all__ = (
     "FnWithKwargs",
     "ListEntry",
     "ModuleEntry",
+    "ModuleLoader",
+    "Storage",
     "extract_variable",
 )