+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.3.0]
+
+### Added
+
+- `Storage` class added
+
+### Changed
+
+- using `ModuleLoader` for loading modules
+
 ## [1.2.1]
 
 ### Fixed